Updated Admin to store simulator instances and use URLs as identifiers for windows and simulator instances
Each simulator is registered in the global namespace (shared by main and renderer processes), to be retrievable in Simulator/preload.js in time. I am not certain as to why registering globally works, as opposed to using Admin.simulators for this retrieval as well. And alternative would be to create the simulator before the browser, and set it to the simulator when ready